+++ internal/config/dsn_test.go
@@ -1683,6 +1683,19 @@ func TestDSN(t *testing.T) {
 		},
 		{
 			cfg: &Config{
+				Account:                             "ac",
+				User:                                "u",
+				Password:                            "p",
+				Database:                            "db",
+				Authenticator:                       AuthTypeWorkloadIdentityFederation,
+				Host:                                "ac.snowflakecomputing.com",
+				WorkloadIdentityProvider:            "aws",
+				WorkloadIdentityAwsUseOutboundToken: BoolTrue,
+			},
+			dsn: "u:p@ac.snowflakecomputing.com:443?account=ac&authenticator=workload_identity&database=db&ocspFailOpen=true&validateDefaultParameters=true&workloadIdentityAwsUseOutboundToken=true&workloadIdentityProvider=aws",
+		},
+		{
+			cfg: &Config{
 				User:                           "u",
 				Password:                       "p",
 				Account:                        "a",